requirements.txt
# =============================================================================
# Genesis AI Deep Agents Service - Dependencies
# =============================================================================

# Core FastAPI & ASGI
fastapi
uvicorn[standard]
python-multipart

# LangGraph & AI Dependencies  
langgraph
langchain
langchain-core
langchain-openai
langchain-anthropic
pydantic[email]

# Database & ORM
sqlalchemy
alembic
asyncpg==0.29.0
aiosqlite

# Redis & Caching
redis[hiredis]
redis-om

# HTTP Client & External APIs
httpx[http2]>=0.23.0
tavily-python
openai
anthropic

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Environment & Configuration
python-dotenv
pydantic-settings

# Logging & Monitoring
structlog
prometheus-client

# Testing
pytest
pytest-asyncio
pytest-mock
faker
asgi-lifespan

# Development
black
isort
flake8
pre-commit
